requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asarPy==1.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Cython==3.0.12
Discode.py==1.1.1
discord.py==2.5.2
Flask==3.1.1
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10.dev0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
multidict==6.4.3
packaging==25.0
propcache==0.3.1
pycairo==1.27.0
Pygments==2.19.1
PyGObject==3.50.0
python-dotenv==1.1.0
requests==2.32.3
setuptools==80.3.1
six==1.17.0
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.46.1
yarl==1.20.0
